// src/app/static/src/generated.d.ts
export interface ModelStatusResponse {
  id: string;
  done: boolean | null;
  status: string;
  success: boolean | null;
  queue: number;
  progress: {
    started: boolean;
    complete: boolean;
    value?: number;
    name: string;
    helpText?: string;
  }[];
}